Default Page Including on index.php
Added the property default_page to the Nav object, and in controller the
get_page_from_uri function will include the default_page if set in the
nav_config.

getter for private default_page property of Nav.

Calling the get_page_from_uri function from the controller on index
page.